+++ packages/custom-clusters/src/clusters/neo.ts
@@ -4,19 +4,19 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { attribute, cluster, single } from "@matter/main/model";
+import { attribute, cluster, uint32 } from "@matter/main/model";
 
 @cluster(0x125dfc11)
 export class NeoCluster {
-    @attribute(0x125d0021, single)
+    @attribute(0x125d0021, uint32)
     wattAccumulated?: number;
 
-    @attribute(0x125d0023, single)
+    @attribute(0x125d0023, uint32)
     watt?: number;
 
-    @attribute(0x125d0022, single)
+    @attribute(0x125d0022, uint32)
     current?: number;
 
-    @attribute(0x125d0024, single)
+    @attribute(0x125d0024, uint32)
     voltage?: number;
 }